refactor(ffmpeg/merge): extract AvPacketOwned RAII wrapper into its own submodule
mkv_raw_ffi.rs (572 lines) carried an inline `AvPacketOwned` struct +
Drop impl + 2 unit tests that have nothing to do with the giant MKV
merge function — they're a generic RAII guard for any heap-allocated
av_packet_alloc/free pair. Move them into a sibling `av_packet_owned.rs`
so they're discoverable by name and reusable across both the standard
(`mod.rs`) and MKV merge paths without either file owning the
definition.

Net file inventory:
  mkv_raw_ffi.rs:      572 -> 517 lines (-55)
  av_packet_owned.rs:    +79 lines (new; 50 LOC + 22 LOC tests + docs)

Why this split (and not a deeper one): the remaining 517 lines of
`mkv_raw_ffi.rs` are a single 459-line `merge_mkv_raw_ffi` function
that drives 11 numbered FFI steps (input open → stream copy → output
write → finalization). Each step touches 5-10 raw pointers + unsafe
blocks + bounds checks. Extracting per-step helpers would require
passing 6+ unsafe arguments across function boundaries and would HURT
readability, not improve it. The merge function is logically one
cohesive FFI sequence and is kept that way.

Sibling helpers stay where they are:
  - raw_ffi_helpers.rs (133): cross-cutting packet read / rescale /
    DTS-compare helpers used by both merge paths. Already separate.

Tests for AvPacketOwned (2) move alongside the type they cover. All
existing FFmpeg merge tests continue to pass.
